add results autosliding and fight view (#12)
