+++ src/app/competitor-zone/referee/referee-zone/referee-zone.component.scss
@@ -25,7 +25,7 @@
         padding-bottom: 0;
         justify-content: center;
         align-items: center;
-        border-bottom: 5px dotted rgba($color-grey-0, 0.3);
+        border-bottom: .2rem dashed rgba($color-grey-0, 0.5);
     }
     &-categories {
         width: 100%;
@@ -35,7 +35,7 @@
         flex-wrap: wrap;
         justify-content: center;
         align-items: center;
-        border-bottom: 5px dotted rgba($color-grey-0, 0.3);
+        border-bottom: .2rem dashed rgba($color-grey-0, 0.5);
         padding: 1rem 0;
         &-header {
             height: 100%;
