+++ src/app/competitor-zone/user/settings/settings.component.scss
@@ -1,5 +1,6 @@
 @import '../../../scss/colors';
 @import '../../../scss/variables';
+@import '../../register/register.component.scss';
 .scroll {
     // height: 50rem;
     display: flex;
@@ -134,6 +135,19 @@ section {
 .top {
     border-top: rgba($color-grey-0, 0.3);
     margin-top: 0rem;
+}
+
+.checkbox-container {
+    display: flex;
+    flex-direction: row;
+    // width: max-content;
+    a {
+        margin-left: 4rem;
+    }
+}
+
+.checkbox-checked {
+    color: white;
 }
 
 @media only screen and (max-width: 1150px) {
