+++ src/app/home/results/results.component.scss
@@ -4,22 +4,24 @@
 @import '../../competitor-zone/referee/referee-zone/referee-zone.component.scss';
 .results {
     width: 100%;
-    height: 100vh;
-    padding-top: 9rem;
+    min-height: 95vh;
+    height: min-content;
+    padding-top: 12rem;
     display: flex;
     flex-direction: column;
-    justify-content: center;
+    // justify-content: center;
     align-items: center;
-    p {
-        font-size: 2rem;
-    }
+    // p {
+    //     font-size: 2rem;
+    // }
     &-main {
         // overflow: scroll;
         display: flex;
         flex-direction: column;
         align-items: center;
-        width: 90%;
-        height: 90%;
+        width: 95%;
+        min-height: 80vh;
+        // height: 90%;
         padding: 3rem;
         box-shadow: 2rem 2rem 4rem $color-grey-4, -2rem -2rem 4rem rgba($color-grey-0, 0.1), inset 2rem 2rem 4rem $color-grey-4, inset -2rem -2rem 4rem rgba($color-grey-0, 0.1);
         border-radius: 4rem;
@@ -28,14 +30,15 @@
 
 .inline {
     min-height: 8rem;
-    overflow-y: hidden;
+    // overflow-y: hidden;
     height: max-content !important;
     width: 100% !important;
     max-width: 100%;
-    flex-wrap: nowrap;
-    overflow-x: auto !important;
+    flex-wrap: wrap;
+    // overflow-x: auto !important;
     display: flex;
-    justify-content: flex-start;
+    justify-content: center;
+    border-bottom: .2rem dashed rgba($color-grey-0, 0.5);
     // align-items: flex-start;
     // overflow-y: hidden !important;
     button {
