+++ src/environments/environment.prod.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
-  apiUrl: 'https://api.xchallenge.pl/'
+  apiUrl: 'http://127.0.0.1:8080/'
 
 };
